chore: Refactored - Add integration Test - Add Unit Tests

+++ spring-expense-tracker/src/main/java/com/harisw/springexpensetracker/application/expense/service/DeleteExpenseService.java
@@ -1,5 +1,6 @@
 package com.harisw.springexpensetracker.application.expense.service;
 
+import com.harisw.springexpensetracker.domain.expense.ExpenseNotFoundException;
 import com.harisw.springexpensetracker.domain.expense.ExpenseRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,6 +17,9 @@ public class DeleteExpenseService {
     }
 
     public void delete(UUID publicId) {
+        if (repository.findByPublicId(publicId).isEmpty()) {
+            throw new ExpenseNotFoundException(publicId);
+        }
         repository.deleteByPublicId(publicId);
     }
 }
